Build src migration: Herding Elephants. (#53)
* Get rid of buildSrc and pre-compiled scripts

* create gradle plugins.

* Switch to plugin scripts and cleanup dependencies.

* Update integers to long

* Update import after upgrading sqlDelight

* Create kmm scripts

* Provide coroutineContext

* Provide coroutineContext

* Use kmm plugins & update dependencies.

* Delete unused class

* Cleanup: Move common block to script

* Cleanup: update gradle scripts.

* Minor cleanup: add manifestPlaceholders, add suppress annotation.

* Add app flavor extension.

* Simplify build job

* Remove unused blank sourceSets

* Remove unused dependency

* Update lib

* Add flow preview to opt-in.

* Update dependencies.

* Add artwork

* Update readme.

+++ android/features/following/src/main/kotlin/com/thomaskioko/tvmaniac/following/FollowingListScreen.kt
@@ -34,7 +34,7 @@ import com.thomaskioko.tvmaniac.resources.R
 @Composable
 fun FollowingScreen(
     viewModel: FollowingViewModel,
-    openShowDetails: (showId: Int) -> Unit,
+    openShowDetails: (showId: Long) -> Unit,
 ) {
 
     val followedState by viewModel.state.collectAsStateWithLifecycle()
@@ -77,7 +77,7 @@ fun FollowingScreen(
 private fun FollowingGridContent(
     list: List<FollowedShow>,
     paddingValues: PaddingValues,
-    onItemClicked: (Int) -> Unit,
+    onItemClicked: (Long) -> Unit,
 ) {
     val listState = rememberLazyListState()
