+++ android/features/shows-grid/src/main/kotlin/com/thomaskioko/tvmaniac/show_grid/GridActions.kt
@@ -2,5 +2,5 @@ package com.thomaskioko.tvmaniac.show_grid
 
 sealed interface GridActions
 
-data class ReloadShows(val category: Int)  : GridActions
-data class LoadShows(val category: Int) : GridActions
+data class ReloadShows(val category: Long)  : GridActions
+data class LoadShows(val category: Long) : GridActions
