+++ shared/domain/episodes/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/episodes/api/EpisodeRepository.kt
@@ -3,5 +3,5 @@ package com.thomaskioko.tvmaniac.episodes.api
 
 interface EpisodeRepository {
 
-    suspend fun updateEpisodeArtWork(showId: Int)
+    suspend fun updateEpisodeArtWork(showId: Long)
 }
