+++ shared/domain/episodes/testing/src/commonMain/kotlin/com/thomaskioko/tvmaniac/episodes/testing/FakeEpisodeRepository.kt
@@ -4,5 +4,5 @@ import com.thomaskioko.tvmaniac.episodes.api.EpisodeRepository
 
 class FakeEpisodeRepository : EpisodeRepository {
 
-    override suspend fun updateEpisodeArtWork(showId: Int) {}
+    override suspend fun updateEpisodeArtWork(showId: Long) {}
 }
