+++ shared/domain/season-details/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/seasondetails/api/model/SeasonDetails.kt
@@ -2,9 +2,9 @@ package com.thomaskioko.tvmaniac.seasondetails.api.model
 
 
 data class SeasonDetails(
-    val seasonId: Int,
+    val seasonId: Long,
     val seasonName: String,
-    val episodeCount: Int,
+    val episodeCount: Long,
     val watchProgress: Float,
     val episodes: List<Episode>
 )
