+++ shared/domain/show-details/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/details/api/model/Show.kt
@@ -1,8 +1,8 @@
 package com.thomaskioko.tvmaniac.details.api.model
 
 data class Show(
-    val traktId: Int = 0,
-    val tmdbId: Int? = 0,
+    val traktId: Long = 0,
+    val tmdbId: Long? = 0,
     val title: String = "",
     val overview: String = "",
     val language: String? = null,
@@ -10,9 +10,9 @@ data class Show(
     val backdropImageUrl: String? = null,
     val year: String = "",
     val status: String? = null,
-    val votes: Int = 0,
+    val votes: Long = 0,
     val numberOfSeasons: Int? = null,
-    val numberOfEpisodes: Int? = null,
+    val numberOfEpisodes: Long? = null,
     val rating: Double = 0.0,
     val genres: List<String> = listOf(),
     val isFollowed: Boolean = false,
