+++ shared/domain/shows/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/shows/api/model/TvShow.kt
@@ -1,8 +1,8 @@
 package com.thomaskioko.tvmaniac.shows.api.model
 
 data class TvShow(
-    val traktId: Int = 0,
-    val tmdbId: Int? = 0,
+    val traktId: Long = 0,
+    val tmdbId: Long? = 0,
     val title: String = "",
     val overview: String = "",
     val language: String? = null,
@@ -10,9 +10,9 @@ data class TvShow(
     val backdropImageUrl: String? = null,
     val year: String = "",
     val status: String? = null,
-    val votes: Int = 0,
-    val numberOfSeasons: Int? = null,
-    val numberOfEpisodes: Int? = null,
+    val votes: Long = 0,
+    val numberOfSeasons: Long? = null,
+    val numberOfEpisodes: Long? = null,
     val rating: Double = 0.0,
     val genres: List<String> = listOf(),
     val isFollowed: Boolean = false,
